backend/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
